flaskweb/flask-backend/requirements.txt
click==8.0.1
cmake==3.20.5
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
Pillow==8.3.1
six==1.16.0
Werkzeug==2.0.1

PyMySQL~=1.0.2